requirements.txt
pulsectl
pyserial-asyncio
asyncio